refactor: change header folder position
