+++ src/components/App.js
@@ -1,14 +1,8 @@
 import garage from "../view/garage";
 import winners from "../view/winners";
-import Header from "../view/header";
+import Header from "./header";
 
 const App = () => {
-
-    if (!localStorage.getItem('firstLoad')) {
-        localStorage.setItem('page', 'garage');
-        localStorage.setItem('firstLoad', ' ');
-    }
-
     document.getElementById('root').innerHTML = Header;
     garage();
     winners();
