--- src/components/header/header.js
@@ -1,6 +1,0 @@
-const Header = `<div class="header-wrapper">
-                    <button class="header-button" id="garage">TO GARAGE</button>
-                    <button class="header-button" id="winners">TO WINNERS</button>
-                </div>`
-
-export default Header;
