+++ src/style.css
@@ -4,7 +4,6 @@ body {
 
 .header-wrapper {
     height: 60px;
-    margin: 40px;
     display: flex;
     align-items: center;
 }
