+++ src/view/winners.js
@@ -1,4 +1,4 @@
-import Header from "./header";
+import Header from "../components/header";
 import load from "../controller/loader";
 
 const winners = () => {
